// public/types.tsx
export interface User {
    authId: string;
    email: string;
    first_name: string;
    history: any[];
    id: string;
    image_url: string;
    last_name: string;
    level: string;
    name: string;
    score: number;
    sub_id: string;
    tel: string;
}

export interface Score {
    level: string;
    score: number;
}


export interface Quiz {
    authId: string;
    email: string;
    first_name: string;
    history: any[];
    id: string;
    image_url: string;
    last_name: string;
    level: string;
    name: string;
    score: number;
    sub_id: string;
    tel: string;
}